CMP320/Scripting/WAVARG/requirements.txt
python-docx
python-nmap
pydirbuster
bs4
selenium
webdriver_manager
wfuzz